Refactor math tools: rename check-math-agentic to check-math-with-mathjs
Major restructuring of math verification tools:

## Tool Renaming
- check-math-with-mathjs → check-math-old (deprecated legacy tool)
- check-math-agentic → check-math-with-mathjs (new primary agentic tool)

## Schema Updates
- Updated provide_verdict tool schema to match original output format
- Restored verificationDetails with mathJsExpression, computedValue, steps
- Maintained errorDetails structure from original tool
- Added error field for technical failures

## Integration Updates
- Updated tool registry to reference renamed tools
- Updated API routes and UI pages
- Updated check-math-hybrid to use new check-math-with-mathjs
- Updated math analysis plugin imports
- Updated all test files with new naming and schema expectations

## Benefits
- Simplified 2-tool agent architecture (evaluate_expression + provide_verdict)
- Maintains familiar output schema for backwards compatibility
- Enables prompt caching (under 4-tool Anthropic limit)
- Preserves all existing functionality while improving efficiency

The new check-math-with-mathjs combines the best of both approaches: the powerful agentic architecture with the familiar output format.
