+++ src/app/api/tools/check-math-old/route.ts
@@ -1,0 +1,4 @@
+import { createToolRoute } from '@/tools/base/createToolRoute';
+import CheckMathOldTool from '@/tools/check-math-old';
+
+export const POST = createToolRoute(CheckMathOldTool);
