+++ src/tools/check-math-old/prompts.ts
