+++ src/tools/check-math-with-mathjs/index.integration.test.ts
@@ -1,220 +1,170 @@
 import { checkMathWithMathJsTool } from './index';
-import { logger } from '@/lib/logger';
+import { ToolContext } from '../base/Tool';
+
+// Mock logger that matches the Logger interface
+const createMockLogger = () => ({
+  info: jest.fn(),
+  error: jest.fn(),
+  warn: jest.fn(),
+  debug: jest.fn(),
+  isDevelopment: false,
+  log: jest.fn(),
+  logRequest: jest.fn(),
+  logResponse: jest.fn(),
+  child: jest.fn(() => createMockLogger())
+});
 
 describe('CheckMathWithMathJsTool Integration Tests', () => {
-  const mockContext = {
-    logger: {
-      info: jest.fn(),
-      error: jest.fn(),
-      warn: jest.fn(),
-      debug: jest.fn()
-    }
-  };
+  let mockContext: ToolContext;
 
   beforeEach(() => {
-    jest.clearAllMocks();
+    mockContext = {
+      logger: createMockLogger()
+    };
   });
 
-  describe('Simple expressions that should use direct MathJS evaluation', () => {
-    it('should verify simple arithmetic quickly', async () => {
-      const startTime = Date.now();
+  describe('Early symbolic math detection', () => {
+    it('should quickly return cannot_verify for derivative statements', async () => {
       const result = await checkMathWithMathJsTool.execute({
-        statement: '2 + 2 = 4'
+        statement: "The derivative of x³ is 3x²"
       }, mockContext);
-      const duration = Date.now() - startTime;
-      
-      expect(result.status).toBe('verified_true');
-      expect(duration).toBeLessThan(100); // Should be nearly instant
-      expect(result.verificationDetails?.mathJsExpression).toBe('2 + 2 == 4');
+
+      expect(result.status).toBe('cannot_verify');
+      expect(result.explanation).toBe('Cannot verify symbolic math. MathJS only handles numerical computations.');
+      expect(result.llmInteraction.tokensUsed.total).toBe(0); // No LLM call made
     });
 
-    it('should catch simple arithmetic errors', async () => {
+    it('should detect various symbolic math keywords', async () => {
+      const symbolicStatements = [
+        "The integral of sin(x) is -cos(x)",
+        "Prove that (a+b)² = a² + 2ab + b²",
+        "The limit as x approaches 0 of sin(x)/x is 1",
+        "Simplify the expression x² + 2x + 1",
+        "d/dx(e^x) = e^x"
+      ];
+
+      for (const statement of symbolicStatements) {
+        const result = await checkMathWithMathJsTool.execute({ statement }, mockContext);
+        expect(result.status).toBe('cannot_verify');
+        expect(result.explanation).toBe('Cannot verify symbolic math. MathJS only handles numerical computations.');
+      }
+    });
+  });
+
+  describe('Numerical verification with concise explanations', () => {
+    it('should verify simple arithmetic with brief explanation', async () => {
       const result = await checkMathWithMathJsTool.execute({
-        statement: '2 + 2 = 5'
+        statement: "2 + 2 = 4"
       }, mockContext);
-      
+
+      expect(result.status).toBe('verified_true');
+      expect(result.explanation.length).toBeLessThan(100); // Ensure explanation is concise
+      expect(result.verificationDetails?.mathJsExpression).toBeTruthy();
+    }, 30000);
+
+    it('should detect errors with brief explanation', async () => {
+      const result = await checkMathWithMathJsTool.execute({
+        statement: "10% of 50 is 10"
+      }, mockContext);
+
       expect(result.status).toBe('verified_false');
-      expect(result.errorDetails?.conciseCorrection).toBe('5 → 4');
-      expect(result.errorDetails?.errorType).toBe('calculation');
-    });
+      expect(result.explanation).toContain('5');
+      expect(result.explanation.length).toBeLessThan(100);
+      expect(result.errorDetails).toBeDefined();
+      expect(result.errorDetails?.conciseCorrection).toBe('10 → 5');
+    }, 30000);
 
-    it('should handle percentage calculations', async () => {
+    it('should handle unit conversions with rounding tolerance', async () => {
       const result = await checkMathWithMathJsTool.execute({
-        statement: '30% of 150 is 45'
+        statement: "Converting 100 fahrenheit to celsius gives 37.78 degrees"
       }, mockContext);
-      
-      expect(result.status).toBe('verified_true');
-    });
 
-    it('should handle unit conversions', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: '5 kilometers equals 5000 meters'
-      }, mockContext);
-      
+      // Should accept reasonable rounding
       expect(result.status).toBe('verified_true');
-    });
+      expect(result.explanation).toContain('rounding');
+    }, 30000);
+
+    it('should handle percentages correctly', async () => {
+      const result = await checkMathWithMathJsTool.execute({
+        statement: "30% equals 0.3"
+      }, mockContext);
+
+      expect(result.status).toBe('verified_true');
+      expect(result.verificationDetails?.mathJsExpression).toBeTruthy();
+    }, 30000);
   });
 
-  describe('Complex expressions requiring LLM extraction', () => {
-    it('should handle statistical statements', async () => {
-      const startTime = Date.now();
+  describe('Tool calls tracking', () => {
+    it('should track tool calls made during verification', async () => {
       const result = await checkMathWithMathJsTool.execute({
-        statement: 'The risk of death in 2019 was 0.00202% per person-day'
+        statement: "5 factorial equals 120"
       }, mockContext);
-      const duration = Date.now() - startTime;
-      
-      // This should go through extraction first, then maybe full LLM
-      expect(duration).toBeLessThan(5000); // Should be under 5 seconds with Haiku
-      expect(['verified_true', 'cannot_verify']).toContain(result.status);
-    });
 
-    it('should handle scientific notation', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: '3.39*10^-7 equals 0.000000339'
-      }, mockContext);
-      
-      expect(result.status).toBe('verified_true');
-    });
-
-    it('should handle word problems', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: 'If I have twenty apples and give away twelve, I have eight left',
-        context: 'Word problem with numbers written as words'
-      }, mockContext);
-      
-      expect(result.status).toBe('verified_true');
-    });
+      expect(result.verificationDetails?.mathJsExpression).toBeTruthy();
+      expect(result.reasoning.length).toBeGreaterThan(50); // Should have detailed reasoning
+    }, 30000);
   });
 
-  describe('Edge cases and error scenarios', () => {
+  describe('Complex expressions', () => {
+    it('should handle binomial coefficients', async () => {
+      const result = await checkMathWithMathJsTool.execute({
+        statement: 'The binomial coefficient "10 choose 3" equals 120'
+      }, mockContext);
+
+      expect(result.status).toBe('verified_true');
+      expect(result.verificationDetails?.mathJsExpression).toBeTruthy();
+    }, 30000);
+
+    it('should handle unit operations', async () => {
+      const result = await checkMathWithMathJsTool.execute({
+        statement: "5 kilometers plus 3000 meters equals 8 kilometers"
+      }, mockContext);
+
+      expect(result.status).toBe('verified_true');
+    }, 30000);
+  });
+
+  describe('Error handling', () => {
     it('should handle malformed expressions gracefully', async () => {
       const result = await checkMathWithMathJsTool.execute({
-        statement: '2 + + 3 = 5'
+        statement: "2 + + 3 = 5"
       }, mockContext);
-      
+
+      expect(['verified_false', 'cannot_verify']).toContain(result.status);
+      expect(result.explanation).toBeTruthy();
+    }, 30000);
+
+    it('should handle incomplete statements', async () => {
+      const result = await checkMathWithMathJsTool.execute({
+        statement: "0.736 % ... So the risk is 0.00202 %"
+      }, mockContext);
+
       expect(result.status).toBe('cannot_verify');
-      expect(result.explanation).toContain('verify');
-    });
-
-    it('should handle non-mathematical statements', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: 'The sky is blue'
-      }, mockContext);
-      
-      expect(result.status).toBe('cannot_verify');
-    });
-
-    it('should handle very long mathematical expressions', async () => {
-      const longExpression = '1 + 2 + 3 + 4 + 5 + 6 + 7 + 8 + 9 + 10 + ' +
-                           '11 + 12 + 13 + 14 + 15 + 16 + 17 + 18 + 19 + 20 = 210';
-      
-      const result = await checkMathWithMathJsTool.execute({
-        statement: longExpression
-      }, mockContext);
-      
-      expect(result.status).toBe('verified_true');
-    });
-
-    it('should handle expressions with special characters', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: '√16 + π ≈ 7.14'
-      }, mockContext);
-      
-      // Should handle special math symbols
-      expect(['verified_true', 'verified_false']).toContain(result.status);
-    });
+      expect(result.explanation).toContain('incomplete');
+    }, 30000);
   });
 
-  describe('Performance under load', () => {
-    it('should handle multiple expressions in sequence efficiently', async () => {
-      const expressions = [
-        '2 + 2 = 4',
-        '10 * 5 = 50',
-        '100 / 4 = 25',
-        '7 - 3 = 4',
-        '2^8 = 256'
-      ];
-      
-      const startTime = Date.now();
-      const results = await Promise.all(
-        expressions.map(statement => 
-          checkMathWithMathJsTool.execute({ statement }, mockContext)
-        )
-      );
-      const totalDuration = Date.now() - startTime;
-      
-      expect(results).toHaveLength(5);
-      expect(results.every(r => r.status === 'verified_true')).toBe(true);
-      expect(totalDuration).toBeLessThan(1000); // All should complete in under 1 second
-    });
-
-    it('should handle a mix of simple and complex expressions', async () => {
-      const expressions = [
-        { statement: '2 + 2 = 4', expected: 'verified_true' },
-        { statement: '30% of 150 is 45', expected: 'verified_true' },
-        { statement: 'The probability is 0.00202%', expected: 'cannot_verify' },
-        { statement: '5 km = 5000 m', expected: 'verified_true' },
-        { statement: '3.14159 ≈ π', expected: 'verified_true' }
-      ];
-      
-      const startTime = Date.now();
-      const results = await Promise.all(
-        expressions.map(({ statement }) => 
-          checkMathWithMathJsTool.execute({ statement }, mockContext)
-        )
-      );
-      const totalDuration = Date.now() - startTime;
-      
-      expect(results).toHaveLength(5);
-      expect(totalDuration).toBeLessThan(10000); // Should complete in under 10 seconds total
-    });
-  });
-
-  describe('LLM extraction fallback', () => {
-    it('should extract and evaluate complex equality', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: 'The sum of the first 10 natural numbers equals 55'
+  describe('Session management', () => {
+    it('should create unique sessions for each execution', async () => {
+      // Execute twice and check logs for different session IDs
+      await checkMathWithMathJsTool.execute({
+        statement: "1 + 1 = 2"
       }, mockContext);
       
-      // Should extract "1+2+3+4+5+6+7+8+9+10 == 55" or similar
-      expect(result.status).toBe('verified_true');
-      if (result.verificationDetails?.mathJsExpression) {
-        expect(result.verificationDetails.computedValue).toContain('true');
-      }
-    });
-
-    it('should handle extraction retries for ambiguous statements', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: 'Half of sixty plus twenty equals fifty'
+      const firstSessionCall = (mockContext.logger.info as jest.Mock).mock.calls
+        .find(call => call[0].includes('Created new session'));
+      
+      await checkMathWithMathJsTool.execute({
+        statement: "2 + 2 = 4"
       }, mockContext);
       
-      // Could be interpreted as (60/2) + 20 = 50 or 60/2 + 20 = 50
-      expect(['verified_true', 'verified_false']).toContain(result.status);
-    });
-  });
-
-  describe('Error details and corrections', () => {
-    it('should provide clear corrections for calculation errors', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: '15 * 7 = 100'
-      }, mockContext);
+      const secondSessionCall = (mockContext.logger.info as jest.Mock).mock.calls
+        .filter(call => call[0].includes('Created new session'))
+        .pop();
       
-      expect(result.status).toBe('verified_false');
-      expect(result.errorDetails?.conciseCorrection).toBe('100 → 105');
-      expect(result.errorDetails?.errorType).toBe('calculation');
-      expect(result.errorDetails?.severity).toBe('major');
-    });
-
-    it('should handle rounding errors appropriately', async () => {
-      const result = await checkMathWithMathJsTool.execute({
-        statement: '1/3 = 0.333'
-      }, mockContext);
-      
-      // Should recognize this as approximately correct or a rounding issue
-      expect(['verified_true', 'verified_false']).toContain(result.status);
-      if (result.status === 'verified_false') {
-        expect(result.errorDetails?.errorType).toBe('calculation');
-      }
-    });
+      expect(firstSessionCall).toBeTruthy();
+      expect(secondSessionCall).toBeTruthy();
+      expect(firstSessionCall[0]).not.toBe(secondSessionCall[0]);
+    }, 30000);
   });
 });
