+++ src/tools/registry.ts
@@ -2,9 +2,9 @@ import { Tool } from './base/Tool';
 import ForecasterTool from './forecaster';
 import FactCheckerTool from './fact-checker';
 import CheckMathTool from './check-math';
+import CheckMathOldTool from './check-math-old';
 import CheckMathWithMathJsTool from './check-math-with-mathjs';
 import CheckMathHybridTool from './check-math-hybrid';
-import CheckMathAgenticTool from './check-math-agentic';
 import ExtractForecastingClaimsTool from './extract-forecasting-claims';
 import PerplexityResearchTool from './perplexity-research';
 import ExtractFactualClaimsTool from './extract-factual-claims';
@@ -22,9 +22,9 @@ export class ToolRegistry {
     this.register(ForecasterTool);
     this.register(FactCheckerTool);
     this.register(CheckMathTool);
+    this.register(CheckMathOldTool);
     this.register(CheckMathWithMathJsTool);
     this.register(CheckMathHybridTool);
-    this.register(CheckMathAgenticTool);
     this.register(ExtractForecastingClaimsTool);
     this.register(PerplexityResearchTool);
     this.register(ExtractFactualClaimsTool);
